requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
python-docx>=1.1.0
PyYAML>=6.0.1
pdfplumber>=0.10.0
